use correct storage key, documentation, explicit disable of idle animation
