+++ packages/stage-ui-live2d/src/composables/live2d/index.ts
@@ -1,6 +1,6 @@
 export * from './animation'
 export * from './beat-sync'
-export * from './cursor-tracking'
 export * from './expression-controller'
+export * from './eye-tracking'
 export * from './live2d'
 export * from './motion-manager'
